Time: 957 ms (42.03%), Space: 112.2 MB (60.78%) - LeetHub

+++ 315-count-of-smaller-numbers-after-self/315-count-of-smaller-numbers-after-self.cpp
@@ -1,7 +1,7 @@
 #include <ext/pb_ds/assoc_container.hpp> // Common file
 #include <ext/pb_ds/tree_policy.hpp> // Including tree_order_statistics_node_update
 using namespace __gnu_pbds;
-typedef tree<long long, null_type, less_equal<>, rb_tree_tag, tree_order_statistics_node_update> indexed_multiset;
+typedef tree<int, null_type, less_equal<>, rb_tree_tag, tree_order_statistics_node_update> indexed_multiset;
 class Solution {
 public:
     vector<int> countSmaller(vector<int>& nums) {
@@ -12,8 +12,6 @@ public:
             s.erase(s.find_by_order(s.order_of_key(nums[i])));
             ans.push_back(s.order_of_key(nums[i]));
         }
-        // s.erase(s.find_by_order(5));
-        // for(auto x : s)cout<<x<<" ";
         return ans;
     }
 };
